updated SRS interface
