+++ _pkgdown.yml
@@ -76,13 +76,8 @@ reference:
   - transform_xy
 - subtitle: Spatial reference systems
 - contents:
-  - epsg_to_wkt
-  - srs_find_epsg
-  - srs_get_name
-  - srs_is_geographic
-  - srs_is_projected
-  - srs_is_same
-  - srs_to_wkt
+  - srs_convert
+  - srs_query
   - transform_bounds
 - subtitle: Geometry
 - contents:
