+++ man/srs_convert.Rd
@@ -1,29 +1,43 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/RcppExports.R
-\name{srs_to_wkt}
+\name{srs_convert}
+\alias{srs_convert}
+\alias{epsg_to_wkt}
 \alias{srs_to_wkt}
-\title{Convert various spatial reference formats to Well Known Text}
+\title{Convert spatial reference definitions to OGC Well Known Text}
 \usage{
+epsg_to_wkt(epsg, pretty = FALSE)
+
 srs_to_wkt(srs, pretty = FALSE)
 }
 \arguments{
-\item{srs}{Character string containing an SRS definition in various
-formats (see Details).}
+\item{epsg}{Integer EPSG code.}
 
 \item{pretty}{Logical. \code{TRUE} to return a nicely formatted WKT string
 for display to a person. \code{FALSE} for a regular WKT string (the default).}
+
+\item{srs}{Character string containing an SRS definition in various
+formats (see Details).}
 }
 \value{
 Character string containing OGC WKT.
 }
 \description{
+These functions convert various spatial reference formats to Well Known
+Text (WKT).
+}
+\details{
+\code{epsg_to_wkt()} exports the spatial reference for an EPSG code to
+WKT format.
+Wrapper for \code{OSRImportFromEPSG()} in the GDAL Spatial Reference System API
+with output to WKT.
+
 \code{srs_to_wkt()} converts a spatial reference system (SRS) definition
 in various text formats to WKT. The function will examine the input SRS,
 try to deduce the format, and then export it to WKT.
-}
-\details{
-This is a wrapper for \code{OSRSetFromUserInput()} in the GDAL Spatial
-Reference System C API with output to WKT.
+Wrapper for \code{OSRSetFromUserInput()} in the GDAL Spatial Reference System
+API with output to WKT.
+
 The input SRS may take the following forms:
 \itemize{
 \item WKT - to convert WKT versions (see below)
@@ -37,7 +51,7 @@ The input SRS may take the following forms:
 \item PROJJSON (PROJ >= 6.2)
 }
 
-This function is intended to be flexible, but by its nature it is
+\code{srs_to_wkt()} is intended to be flexible, but by its nature it is
 imprecise as it must guess information about the format intended.
 \code{\link[=epsg_to_wkt]{epsg_to_wkt()}} could be used instead for EPSG codes.
 
@@ -54,6 +68,9 @@ WKT1_GDAL. WKT2 will default to the latest revision implemented
 WKT2_2018 since GDAL 3.2
 }
 \examples{
+epsg_to_wkt(5070)
+writeLines(epsg_to_wkt(5070, pretty=TRUE))
+
 srs_to_wkt("NAD83")
 writeLines(srs_to_wkt("NAD83", pretty=TRUE))
 set_config_option("OSR_WKT_FORMAT", "WKT2")
@@ -61,5 +78,5 @@ writeLines(srs_to_wkt("NAD83", pretty=TRUE))
 set_config_option("OSR_WKT_FORMAT", "")
 }
 \seealso{
-\code{\link[=epsg_to_wkt]{epsg_to_wkt()}}
+\link{srs_query}
 }
