--- man/srs_get_name.Rd
@@ -1,28 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R
-\name{srs_get_name}
-\alias{srs_get_name}
-\title{Return the spatial reference system name}
-\usage{
-srs_get_name(srs)
-}
-\arguments{
-\item{srs}{Character string containing an SRS definition in various
-formats (e.g., WKT, PROJ.4 string, well known name such as NAD27, NAD83,
-WGS84, etc).}
-}
-\value{
-Character string containing the name, or empty string.
-}
-\description{
-\code{srs_get_name()} accepts a spatial reference system definition
-in various text formats and returns the name.
-See \code{\link[=srs_to_wkt]{srs_to_wkt()}} for a description of the possible input formats.
-Wrapper of \code{OSRGetName()} in the GDAL Spatial Reference System API.
-}
-\examples{
-srs_get_name("EPSG:5070")
-}
-\seealso{
-\code{\link[=epsg_to_wkt]{epsg_to_wkt()}}, \code{\link[=srs_find_epsg]{srs_find_epsg()}}, \code{\link[=srs_to_wkt]{srs_to_wkt()}}
-}
